update: Initial add of filters and smalls adjustment of folder organization
