+++ src/modules/technician/constants/index.ts
@@ -1,0 +1,13 @@
+export enum technicianStatusEnum {
+  AVAILABLE = 1,
+  OCCUPIED = 2,
+  ATTENDIND = 3,
+  OFF = 4,
+  EXONERATED = 5,
+}
+
+export enum technicianPositionEnum {
+  FIELD = 1,
+  LAB = 2,
+  FIELD_LAB = 3,
+}
