+++ src/modules/technician/dtos/TechnicianResponseDTO.ts
@@ -1,4 +1,5 @@
-import { CreateAddress } from "@shared/types/address.type";
+import { address } from "@shared/types/address.type";
+import { technicianPositionEnum, technicianStatusEnum } from "../constants";
 
 export interface TechnicianResponseDTO {
   id: number;
@@ -7,20 +8,6 @@ export interface TechnicianResponseDTO {
   cpf: string;
   position: technicianPositionEnum;
   status: technicianStatusEnum;
-  address: CreateAddress;
+  address: address;
   updated_at: Date;
-}
-
-export enum technicianStatusEnum {
-  AVAILABLE = 1,
-  OCCUPIED = 2,
-  ATTENDIND = 3,
-  OFF = 4,
-  EXONERATED = 5,
-}
-
-export enum technicianPositionEnum {
-  FIELD = 1,
-  LAB = 2,
-  FIELD_LAB = 3,
 }
